Сетка плиток и итог прогона в диагностике
- Плитки страницы «Производительность» переложены в UniformGrid, число колонок считает TileColumnsConverter по ширине и масштабу шрифта
- Карточка «Сейчас идёт» заменена шестой плиткой: идущая операция, итог последнего прогона либо «нет прогонов»
- Заведён PerformanceRunTracker, итог кладут ScanViewModel и SyncSessionViewModel на успешной ветке
- Подписи и подсказки плиток переведены с многоточия на перенос
- Формат длительности прогона вынесен в PerformanceFormat.Elapsed

+++ SpaceSnoop.Wpf/Bootstrap/Plural.cs
@@ -7,7 +7,7 @@ public static class Plural
         return $"{count} {Word(count, one, few, many)}";
     }
 
-    public static string Word(int count, string one, string few, string many)
+    public static string Word(long count, string one, string few, string many)
     {
         var tail = Math.Abs(count) % 100;
